# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Elite Tenancy API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Listing {
  id: number;
  title: string;
  slug: string;
  description: string;
  price: number;
  pricePeriod: string;
  category: string;
  city: string;
  postcode: string;
  addressLine1?: string;
  bedrooms: number;
  bathrooms: number;
  /** @nullable */
  floorAreaSqm?: number | null;
  furnished: boolean;
  petsAllowed: boolean;
  billsIncluded: boolean;
  status: string;
  isFeatured: boolean;
  isPremium: boolean;
  photos: string[];
  /** @nullable */
  aiMatchScore?: number | null;
  /** @nullable */
  availableFrom?: string | null;
  viewCount: number;
  enquiryCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface ListingInput {
  title: string;
  description: string;
  price: number;
  pricePeriod: string;
  category: string;
  city: string;
  postcode: string;
  addressLine1?: string;
  bedrooms: number;
  bathrooms: number;
  furnished?: boolean;
  petsAllowed?: boolean;
  billsIncluded?: boolean;
  photos?: string[];
  availableFrom?: string;
}

export interface PlatformStats {
  totalListings: number;
  totalLandlords: number;
  totalTenants: number;
  averageLetTime: number;
  citiesCovered: number;
  satisfactionRate: number;
}

export interface BlogArticle {
  id: number;
  title: string;
  slug: string;
  excerpt: string;
  content: string;
  category: string;
  author: string;
  /** @nullable */
  imageUrl?: string | null;
  readTimeMinutes: number;
  tags: string[];
  publishedAt: string;
  createdAt: string;
}

export interface Lead {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  message?: string | null;
  /** @nullable */
  listingId?: number | null;
  /** @nullable */
  listingTitle?: string | null;
  status: string;
  createdAt: string;
}

export interface LeadInput {
  name: string;
  email: string;
  phone?: string;
  message?: string;
  listingId?: number;
  listingTitle?: string;
}

export interface ContactInput {
  name: string;
  email: string;
  phone?: string;
  subject?: string;
  message: string;
}

export interface ValuationInput {
  name: string;
  email: string;
  phone?: string;
  address?: string;
  postcode: string;
  propertyType?: string;
  bedrooms?: number;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface LandlordStats {
  activeListings: number;
  totalTenants: number;
  monthlyRevenue: number;
  pendingMaintenance: number;
  newLeads: number;
  occupancyRate: number;
}

export interface TenantRecord {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  propertyAddress: string;
  rentAmount: number;
  leaseStart: string;
  leaseEnd: string;
  status: string;
}

export interface FinanceTransaction {
  id: number;
  date: string;
  description: string;
  amount: number;
  type: string;
  status: string;
}

export interface FinanceSummary {
  totalRevenue: number;
  monthlyRevenue: number;
  pendingPayments: number;
  transactions: FinanceTransaction[];
}

export interface MaintenanceRequest {
  id: number;
  title: string;
  description: string;
  category: string;
  priority: string;
  status: string;
  /** @nullable */
  propertyAddress?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface MaintenanceInput {
  title: string;
  description: string;
  category: string;
  priority: string;
}

export interface TenantStats {
  rentDue: number;
  rentDueDate: string;
  daysUntilRenewal: number;
  openMaintenanceRequests: number;
  documentsCount: number;
}

export interface TenancyDetails {
  propertyAddress: string;
  landlordName: string;
  landlordEmail: string;
  monthlyRent: number;
  depositAmount: number;
  leaseStart: string;
  leaseEnd: string;
  tenancyType: string;
  status: string;
}

export interface RentRecord {
  id: number;
  dueDate: string;
  /** @nullable */
  paidDate?: string | null;
  amount: number;
  status: string;
  reference: string;
}

export interface TenantDocument {
  id: number;
  name: string;
  type: string;
  uploadedAt: string;
  size: string;
  downloadUrl: string;
}

export interface AdminStats {
  totalListings: number;
  totalLeads: number;
  totalUsers: number;
  totalArticles: number;
  newLeadsToday: number;
  activeListings: number;
}

export interface UserRecord {
  id: number;
  name: string;
  email: string;
  role: string;
  createdAt: string;
  /** @nullable */
  lastActiveAt?: string | null;
}

export interface TenantMatchRequest {
  city?: string;
  minBudget?: number;
  maxBudget: number;
  bedrooms?: number;
  furnished?: boolean;
  petsAllowed?: boolean;
  billsIncluded?: boolean;
  moveInDate?: string;
  lifestyle?: string;
  priorities?: string;
}

export interface TenantMatchResult {
  id: number;
  score: number;
  summary: string;
  highlights: string[];
  listing: Listing;
}

export interface TenantMatchResponse {
  matches: TenantMatchResult[];
}

export type GetListingsParams = {
city?: string;
minPrice?: number;
maxPrice?: number;
bedrooms?: number;
category?: string;
furnished?: boolean;
};