# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Elite Tenancy API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /listings:
    get:
      operationId: getListings
      tags: [listings]
      summary: Get all listings
      parameters:
        - in: query
          name: city
          schema: { type: string }
        - in: query
          name: minPrice
          schema: { type: integer }
        - in: query
          name: maxPrice
          schema: { type: integer }
        - in: query
          name: bedrooms
          schema: { type: integer }
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: furnished
          schema: { type: boolean }
      responses:
        "200":
          description: List of listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Listing"
    post:
      operationId: createListing
      tags: [listings]
      summary: Create a listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListingInput"
      responses:
        "201":
          description: Created listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
...
  /listings/{id}:
    get:
      operationId: getListingById
      tags: [listings]
      summary: Get a listing by ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "404":
          description: Not found
...
  /blog/{slug}:
    get:
      operationId: getBlogArticleBySlug
      tags: [blog]
      summary: Get a blog article by slug
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Blog article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogArticle"
        "404":
          description: Not found
  /leads:
    post:
      operationId: submitLead
      tags: [contact]
      summary: Submit an enquiry lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Lead created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /contact:
    post:
      operationId: submitContact
      tags: [contact]
      summary: Submit a contact form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "200":
          description: Contact submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /valuation:
    post:
      operationId: submitValuation
      tags: [contact]
      summary: Submit a free valuation request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValuationInput"
      responses:
        "200":
          description: Valuation submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /landlord/listings:
    get:
      operationId: getLandlordListings
      tags: [landlord]
      summary: Get landlord's listings
      responses:
        "200":
          description: Landlord listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Listing"
    post:
      operationId: createLandlordListing
      tags: [landlord]
      summary: Create a new listing (landlord)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListingInput"
      responses:
        "201":
          description: Created listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
...
components:
  schemas:
...
    Listing:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        slug: { type: string }
        description: { type: string }
        price: { type: integer }
        pricePeriod: { type: string }
        category: { type: string }
        city: { type: string }
        postcode: { type: string }
        addressLine1: { type: string }
        bedrooms: { type: integer }
        bathrooms: { type: integer }
        floorAreaSqm: { type: ["integer", "null"] }
        furnished: { type: boolean }
        petsAllowed: { type: boolean }
        billsIncluded: { type: boolean }
        status: { type: string }
        isFeatured: { type: boolean }
        isPremium: { type: boolean }
        photos: { type: array, items: { type: string } }
        aiMatchScore: { type: ["integer", "null"] }
        availableFrom: { type: ["string", "null"] }
        viewCount: { type: integer }
        enquiryCount: { type: integer }
        createdAt: { type: string }
        updatedAt: { type: string }
      required:
        - id
        - title
        - slug
        - description
        - price
        - pricePeriod
        - category
        - city
        - postcode
        - bedrooms
        - bathrooms
        - furnished
        - petsAllowed
        - billsIncluded
        - status
        - isFeatured
        - isPremium
        - photos
        - viewCount
        - enquiryCount
        - createdAt
        - updatedAt
    ListingInput:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        price: { type: integer }
        pricePeriod: { type: string }
        category: { type: string }
        city: { type: string }
        postcode: { type: string }
        addressLine1: { type: string }
        bedrooms: { type: integer }
        bathrooms: { type: integer }
        furnished: { type: boolean }
        petsAllowed: { type: boolean }
        billsIncluded: { type: boolean }
        photos: { type: array, items: { type: string } }
        availableFrom: { type: string }
      required:
        - title
        - description
        - price
        - pricePeriod
        - category
        - city
        - postcode
        - bedrooms
        - bathrooms
...
    BlogArticle:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        slug: { type: string }
        excerpt: { type: string }
        content: { type: string }
        category: { type: string }
        author: { type: string }
        imageUrl: { type: ["string", "null"] }
        readTimeMinutes: { type: integer }
        tags: { type: array, items: { type: string } }
        publishedAt: { type: string }
        createdAt: { type: string }
      required:
        - id
        - title
        - slug
        - excerpt
        - content
        - category
        - author
        - readTimeMinutes
        - tags
        - publishedAt
        - createdAt
    Lead:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        phone: { type: ["string", "null"] }
        message: { type: ["string", "null"] }
        listingId: { type: ["integer", "null"] }
        listingTitle: { type: ["string", "null"] }
        status: { type: string }
        createdAt: { type: string }
      required:
        - id
        - name
        - email
        - status
        - createdAt
    LeadInput:
      type: object
      properties:
        name: { type: string }
        email: { type: string }
        phone: { type: string }
        message: { type: string }
        listingId: { type: integer }
        listingTitle: { type: string }
      required:
        - name
        - email
    ContactInput:
      type: object
      properties:
        name: { type: string }
        email: { type: string }
        phone: { type: string }
        subject: { type: string }
        message: { type: string }
      required:
        - name
        - email
        - message
    ValuationInput:
      type: object
      properties:
        name: { type: string }
        email: { type: string }
        phone: { type: string }
        address: { type: string }
        postcode: { type: string }
        propertyType: { type: string }
        bedrooms: { type: integer }
        message: { type: string }
      required:
        - name
        - email
        - postcode
    SuccessResponse:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
      required:
        - success
        - message
...